database/requirements.txt
# 题库转换工具依赖包

pandas>=2.0.0
openpyxl>=3.1.0
python-docx>=1.0.0
